Refactor air-to-aie codebase by moving util funcs to a separate file; initial channel scheduler impl with loop hierarchy awareness
